@jbduncan 2024-03-06
Actually, `slices.ROC` has no benchmarks, so there's no need to run anything here.