@jbduncan 2024-03-11
As discussed in https://github.com/gonum/gonum/pull/1956#issuecomment-1988326786, I'll make some new benchmarks as soon as I can.